refactor: db, index, router, client, user model
