+++ server/db.ts
@@ -1,11 +1,10 @@
-const mongoose = require('mongoose');
-const DB_PORT = process.env.DB_PORT || 27017;
+import mongoose from "mongoose";
+
+const DB_PORT = Number(process.env.DB_PORT) || 27017;
 const DB_NAME = process.env.DB_NAME || 'SpendTune';
 
-mongoose.connect(
-  `mongodb://localhost:${DB_PORT}/${DB_NAME}`
-  )
+mongoose.connect(`mongodb://localhost:${DB_PORT}/${DB_NAME}`)
   .then(() => console.log(`🦆 Database for SpendTune connected @ port ${DB_PORT}!`))
   .catch((err) => console.log(`😞 Sorry, something went wrong! ${err}`));
 
-module.exports = mongoose;
+export default mongoose;
