+++ server/index.ts
@@ -1,19 +1,20 @@
-require('dotenv').config();
-const express = require('express');
+import * as dotenv from 'dotenv';
+import express from 'express';
 // const session = require("express-session"); EXTRA CREDTIS :D
-const cors = require('cors');
-const router = require('./router');
-const app = express();
+import cors from 'cors';
+import router from './router';
 
+dotenv.config()
 const SERVER_PORT = process.env.SERVER_PORT || 3001;
 const SECRET = process.env.SECRET || 'this is not very secure';
-
 
 const corsConfig = {
   origin: 'http://localhost:3000',
   credentials: true,
 };
 
+
+const app = express();
 app.use(cors(corsConfig));
 app.use(express.json());
 // app.use(                     EXTRA CREDTIS :D => authentication
@@ -32,11 +33,7 @@ app.use(express.json());
 // );
 app.use(router);
 
-app.listen(SERVER_PORT, (err) => {
-  if (err) {
-    console.log(`😞 Sorry, something went wrong! ${err}`);
-  } else {
-    console.log(`💰SpendTune💰  server is listening on port ${SERVER_PORT}!`);
-  }
+app.listen(SERVER_PORT, () => {
+  console.log(`💰SpendTune💰  server is listening on port ${SERVER_PORT}!`);
 });
