+++ server/router.ts
@@ -1,5 +1,7 @@
-const router = require('express').Router();
-const controller = require('./controllers/master');
+import { Response, Request, Router } from 'express';
+import controller from './controllers/master';
+
+const router = Router();
 
 // ROUTES FOR API INTERACTION
 router.get('/api/create_link_token', controller.createLinkToken);
@@ -14,10 +16,8 @@ router.post('/logout', controller.logout);
 router.post('/add_category', controller.addCategory);
 router.post('/assign_category', controller.assignCategory);
 
-router.get('*', (res) => {
-  res
-    .status(404)
-    .send('Sorry, not found 😞');
+router.get('*', (req:Request, res:Response) => {
+  res.status(404).send('Sorry, not found 😞');
 });
 
-module.exports = router;
+export default router;
